Tests/LibWeb/Text/input/DOM/DOMImplementation-createHTMLDocument.html

<body>
<script src="../include.js"></script>
<script>
    test(() => {
        println(document);
        println(document instanceof HTMLDocument);
        let newDocument = document.implementation.createHTMLDocument();
        println(newDocument);
        println(newDocument instanceof HTMLDocument);
    });
</script>
</body>
