Tests/LibWeb/Text/expected/DOM/DOMImplementation-createHTMLDocument.txt

[object HTMLDocument]
true
[object HTMLDocument]
true
